requirements.txt
accelerate
git+https://github.com/openai/CLIP.git
numpy
pandas
Pillow
PyYAML
requests
tqdm
urllib3
